requirements.txt

tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.2
tensorflow-estimator==2.8.0
tensorflow-io-gcs-filesystem==0.26.0
jieba==0.42.1

jsonpickle==2.2.0

jupyter-sql==0.1.0

keras==2.8.0
Keras-Preprocessing==1.1.2

pandas==1.4.2
Pillow==9.2.0
pinyin==0.4.0
pipreqs==0.4.11
plotly==5.10.0
ply==3.11 
 
 
beautifulsoup4==4.11.1
 

click==8.1.3 
conda==4.14.0 
cssutils==2.6.0
cufflinks==0.17.3 
fonttools==4.25.0
frozenlist==1.3.1 
gensim==4.3.0
geonamescache==1.5.0
gglsbl==0+unknown
gglsbl3==0.1.4
google-api-core==2.11.0
google-api-python-client==2.48.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
Google-Images-Search==1.4.6
google-pasta==0.2.0
googleapis-common-protos==1.57.0
greenlet==1.1.3.post0
grpcio==1.53.0
guess-language-spirit==0.5.3
h11==0.14.0
h5py==3.7.0
helium==3.1.0
homoglyphs==2.0.4
httplib2==0.21.0
idna==2.10
imageio==2.23.0
importlib-metadata==4.12.0
IPy==1.1 
ipython-sql==0.4.1 
langdetect==1.0.9
libclang==14.0.6
lxml==4.9.3 
nbimporter==0.3.4 
networkx==2.8.8
nltk==3.7 

numpy-financial==1.0.0
oauthlib==3.2.0
opencv-python==4.6.0.66
openpyxl==3.0.10
opt-einsum==3.3.0
outcome==1.2.0 
paddle-bfloat==0.1.7
paddlepaddle==2.4.1
palettable==3.3.0 
prettytable==0.7.2

protobuf==3.19.5

pyasn1==0.4.8
pyasn1-modules==0.2.8
pycld3==0.22
pycosat==0.6.3
 
python-dotenv==1.0.0
python-nmap==0.7.1
python-resize-image==1.1.20
python-whois==0.8.0
python3-protobuf==2.5.0

PyWavelets==1.4.1 
scikit-image==0.19.3
scikit-learn==1.0.2

 
selenium==3.141.0 
 
tld==0.12.6
tldextract==3.3.1 
uritemplate==4.1.1 
virustotal-python==1.0.0 
webdriver-manager==4.0.0
webdrivermanager==0.10.0
webencodings==0.5.1 
xgboost==1.6.2
xlrd==2.0.1
xpinyin==0.7.6
 
